add button back to starkpilled
